Move kernel stuff into separate folder
