+++ src/client/datascience/jupyter/kernels/kernelService.ts
@@ -7,19 +7,19 @@ import { Kernel } from '@jupyterlab/services';
 import * as path from 'path';
 import * as uuid from 'uuid/v4';
 import { CancellationToken } from 'vscode';
-import '../../common/extensions';
-import { traceError, traceInfo, traceWarning } from '../../common/logger';
-import { IFileSystem } from '../../common/platform/types';
-import { IProcessServiceFactory } from '../../common/process/types';
-import { IAsyncDisposableRegistry } from '../../common/types';
-import * as localize from '../../common/utils/localize';
-import { noop } from '../../common/utils/misc';
-import { IInterpreterService, PythonInterpreter } from '../../interpreter/contracts';
-import { captureTelemetry } from '../../telemetry';
-import { JupyterCommands, RegExpValues, Telemetry } from '../constants';
-import { IJupyterExecution, IJupyterKernelSpec, IJupyterSessionManager } from '../types';
-import { JupyterCommandFinder } from './jupyterCommandFinder';
-import { JupyterKernelSpec } from './jupyterKernelSpec';
+import '../../../common/extensions';
+import { traceError, traceInfo, traceWarning } from '../../../common/logger';
+import { IFileSystem } from '../../../common/platform/types';
+import { IProcessServiceFactory } from '../../../common/process/types';
+import { IAsyncDisposableRegistry } from '../../../common/types';
+import * as localize from '../../../common/utils/localize';
+import { noop } from '../../../common/utils/misc';
+import { IInterpreterService, PythonInterpreter } from '../../../interpreter/contracts';
+import { captureTelemetry } from '../../../telemetry';
+import { JupyterCommands, RegExpValues, Telemetry } from '../../constants';
+import { IJupyterExecution, IJupyterKernelSpec, IJupyterSessionManager } from '../../types';
+import { JupyterCommandFinder } from '../jupyterCommandFinder';
+import { JupyterKernelSpec } from '../jupyterKernelSpec';
 
 /**
  * Responsible for kernel management and the like.
